// src/app/modules/reports/print-bill/models/print-search-result.model.ts
export class PrintSearchResult{
    documentNumber: string;
    documentDate: Date;
    documentType: string;
    amount: number;
    customerName: string;
    billingLevel: string;
    billingLevelCode: string;
    billingBranch: string;
    submissionBranch: string;
    collectionBranch: string;
    msa: string;
    sfx: string;
    rateCard: string
}